Update Lucene to 7.4.0

+++ jochre_search/src/main/java/com/joliciel/jochre/search/JochreIndexSearcher.java
@@ -10,6 +10,7 @@ import org.apache.lucene.search.TopDocs;
 
 /**
  * A reusable index searcher tied to a given index directory.
+ * 
  * @author Assaf Urieli
  *
  */
@@ -18,27 +19,28 @@ public interface JochreIndexSearcher {
 	 * The directory in which the original PDF files are contained.
 	 */
 	public File getContentDir();
-	
+
 	/**
 	 * Return a list of Lucene docIds and scores corresponding to a given query.
 	 */
 	public TopDocs search(JochreQuery query);
-	
+
 	/**
 	 * Find all documents corresponding to a given name.
 	 */
-	public Map<Integer,Document> findDocuments(String name);
-	
+	public Map<Integer, Document> findDocuments(String name);
+
 	/**
 	 * Find the documents for a given name and index.
 	 */
-	public Map<Integer,Document> findDocument(String name, int index);
+	public Map<Integer, Document> findDocument(String name, int index);
 
 	/**
 	 * Write query results in JSON to the provided Writer.
+	 * 
 	 * @return the number of results
 	 */
-	public int search(JochreQuery query, Writer out);
+	public long search(JochreQuery query, Writer out);
 
 	public File getIndexDir();
